Deduplicate inline `fn get_gmv_2d` to `src/env.rs`.
